MetDNA2

MetDNA2(
  ms1_file = "data.csv",
  ms2_file = NULL,
  sample_info_file = "sample.info.csv",
  metdna_version = c("version2", "version1"),
  ms2_type = c("mgf", "mzXML", "msp", "cef"),
  path = ".",
  thread = 3,
  is_check_data = TRUE,
  is_anno_initial_seed = TRUE,
  is_anno_mrn = TRUE,
  is_credential = TRUE,
  is_bio_interpret = TRUE,
  is_exported_report = TRUE,
  is_cred_pg_filter = FALSE,
  is_cred_formula_filter = FALSE,
  lib = c("zhuMetLib", "zhuMetLib_obitrap", "fiehnHilicLib"),
  polarity = c("positive", "negative"),
  instrument = c("SciexTripleTOF", "AgilentQTOF", "BrukerQTOF", "ThermoOrbitrap",
    "Exploris", "AgilentDTIMMS", "BrukerTIMS"),
  column = c("hilic", "rp"),
  ce = c("30", "10", "20", "35,15", "40", "50", "NCE10", "NCE20", "NCE30", "NCE40",
    "NCE50", "SCE20_30_40%", "SNCE20_30_40%"),
  method_lc = c("Amide12min", "Amide23min", "Other"),
  is_rt_calibration = TRUE,
  mz_tol = 25,
  tolerance_rt_range = 30,
  tolerance_ccs_range = 4,
  int_ms2_min_abs = 50,
  dp_cutoff = 0.8,
  direction = c("reverse", "forward"),
  mz_tol_combine_ms1_ms2 = 25,
  rt_tol_combine_ms1_ms2 = 10,
  is_plot_ms2 = TRUE,
  extension_step = c("0", "1", "2", "3", "4", "5", "6", "7", "8"),
  max_isotope = 4,
  rt_tol1 = 3,
  rt_tol2 = 30,
  ccs_tol = 4,
  cor_tol = 0,
  int_tol = 500,
  dp_tol = 0.5,
  max_step = 3,
  score_cutoff = 0,
  remain = FALSE,
  remain_per = 0.5,
  seed_neighbor_match_plot = TRUE,
  candidate_num = 3,
  scoring_approach_recursive = c("dp", "bonanza", "hybrid", "gnps"),
  matched_frag_cutoff = 1,
  test_old_mrn = c("v0.5", "v0.3"),
  test_adduct_version = c("version2", "version1"),
  use_redun_rm_result = FALSE,
 
    dir_GenForm = "/home/zhouzw/software/genform-code-r8-trunk/src/GenForm/GenForm/bin/Release",
  is_pred_formula_all = FALSE,
  formula_mz_tol_ms1 = 25,
  formula_mz_tol_ms2 = 25,
  formula_elements = "CHNOPS",
  formula_candidate_num = 3,
  platform = "linux",
  test_rm_extra_anno_from_ini_seed = FALSE,
  comp_group = c("W03", "W30"),
  uni_test = c("t", "wilcox", "anova"),
  correct_p = FALSE,
  p_cutoff = 0.05,
  species = c("hsa", "mmu", "rno", "bta", "gga", "dre", "dme", "cel", "sce", "osa",
    "ath", "smm", "pfa", "tbr", "eco", "bsu", "ppu", "sau", "tma", "syf", "mln"),
  quanti_pathway_method = c("mean", "sum", "median")
)

Arguments

ms1_file
The name of ms1 data

ms2_file
The names of ms2 data

sample_info_file
'sample.info.csv'

metdna_version
MetDNA parameter 'version1' and 'version2'

ms2_type
'mgf', 'mzXML', 'msp', 'cef' are supported. Default: 'mgf'

path
Default: '.'

thread
The number of threads. Default: 4

is_check_data
Whether check data quality. Default: TRUE

is_anno_initial_seed
Whether annotate initial seeds with spectral DB. Default: TRUE

is_anno_mrn
Whether perform recursive annotation. Default: TRUE

is_credential
Whether perform annotation credential. Default: TRUE

is_cred_formula_filter
Whether filter error annotations by formula prediction in final table. Default: FALSE

lib
Default: 'zhuMetLib'

polarity
'positive', 'negative'. Default: 'positive'

instrument
The instrument you used to acquire data. "SciexTripleTOF", "AgilentQTOF", "BrukerQTOF", "ThermoOrbitrap", 'Exploris', "IMMS". Default: "SciexTripleTOF"

column
"hilic", "rp". Default: 'hilic'

ce
"30", "10", "20", "35,15", "40", "50". Default: "30"

method_lc
'Amide12min', 'Amide23min', 'Other'. Default: "Amide12min"

is_rt_calibration
Only valid with 'Amide12min' and 'Amide23min' methods. Default: TRUE.

mz_tol
The tolerance of ms1 match. Default: 25 ppm

tolerance_rt_range
The maxmium rt tolerance. Default: 30 s

dp_cutoff
MS2 score cutoff. Default: 0.8

direction
Reserve which direct dot product. 'reverse', 'forward'. Default: 'reverse'

is_plot_ms2
Whether plot initial-seed ms2 match plot. Default: TRUE

extension_step
The used extended MRN with n step reaction. Steps: '0', '1', '2', '3', '4', '5', '6', '7', '8'

max_isotope
Maximum considerable isotope. Default: 4

rt_tol1
RT tolerance of adductAnnotation and isotopeAnnotation. Default: 3s

rt_tol2
RT tolerance of metAnnotation. Default: 30

cor_tolDefault: 0

int_tolisotope intensity tolerance. Defautl: 500

dp_tol0.5

max_stepthe maximum step of reaction pair during recursive annotation

score_cutoffTotal score cutoff. Default: 0

remainFALSE

remain_per0.5

seed_neighbor_match_plotTRUE

candidate_numreserve top N candidate number in recursive annotation. Default: 3

is_pred_formula_allwhether predict formula for all peak group (TRUE) or concised peak groups (FALSE). Default: FALSE

formula_mz_tol_ms1Default: 25 ppm

formula_mz_tol_ms2Default: 25 ppm

formula_elementsDefault: "CHNOPS"

formula_candidate_numReserve top 3 formula. Default: 3

platform'windows', 'linux'. Default: 'linux'

is_cred_pg_filerWhether filter error annotations by peak group annotaion in final table. Default: FALSE